requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   orb-billing
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via orb-billing
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via orb-billing
idna==3.10
    # via
    #   anyio
    #   httpx
orb-billing==2.19.0
    # via -r requirements.in
pydantic==2.10.3
    # via orb-billing
pydantic-core==2.27.1
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   orb-billing
typing-extensions==4.12.2
    # via
    #   anyio
    #   orb-billing
    #   pydantic
    #   pydantic-core
